WIP more changes to reflect new sampling API

+++ scanner/util/common.cpp
@@ -83,13 +83,14 @@ std::string image_encoding_type_to_string(ImageEncodingType t) {
   return s;
 }
 
-int PUS_PER_NODE = 1;             // Number of available PUs per node
-int GPUS_PER_NODE = 1;            // Number of available GPUs per node
-i32 IO_ITEM_SIZE = 64;            // Number of rows to load and save at a time
-int WORK_ITEM_SIZE = 8;           // Max size of a work item
-int TASKS_IN_QUEUE_PER_PU = 4;    // How many tasks per PU to allocate to a node
-int LOAD_WORKERS_PER_NODE = 2;    // Number of worker threads loading data
-int SAVE_WORKERS_PER_NODE = 2;    // Number of worker threads loading data
+i32 PUS_PER_NODE = 1;             // Number of available PUs per node
+i32 GPUS_PER_NODE = 1;            // Number of available GPUs per node
+i64 IO_ITEM_SIZE = 64;            // Number of rows to load and save at a time
+i64 WORK_ITEM_SIZE = 8;           // Max size of a work item
+i32 TASKS_IN_QUEUE_PER_PU = 4;    // How many tasks per PU to allocate to a node
+i32 LOAD_WORKERS_PER_NODE = 2;    // Number of worker threads loading data
+i32 SAVE_WORKERS_PER_NODE = 2;    // Number of worker threads loading data
 std::vector<i32> GPU_DEVICE_IDS;  // GPU device ids to use
-int NUM_CUDA_STREAMS = 32;        // Number of cuda streams for image processing
+i32 NUM_CUDA_STREAMS = 32;        // Number of cuda streams for image processing
+
 }
